3ds: it compiles

+++ common/source/tobkit/envelope_editor.h
@@ -61,13 +61,13 @@ class EnvelopeEditor: public Widget
 {
 	public:
 		// Constructor sets base variables
-		EnvelopeEditor(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen, u16 _max_x, u16 _max_y, u16 _max_points);
+		EnvelopeEditor(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen, u16 _max_x, u16 _max_y, u16 _max_points);
 		~EnvelopeEditor(void);
 
 		// Event calls
-		void penDown(u8 px, u8 py);
-		void penUp(u8 px, u8 py);
-		void penMove(u8 px, u8 py);
+		void penDown(u16 px, u16 py);
+		void penUp(u16 px, u16 py);
+		void penMove(u16 px, u16 py);
 
 		// Drawing request
 		void pleaseDraw(void);
@@ -95,7 +95,7 @@ class EnvelopeEditor: public Widget
 		void draw(void);
 
 		void calcScrollThingy(void);
-		void drawPoint(u8 x, u8 y, bool active);
+		void drawPoint(u16 x, u16 y, bool active);
 
 		void scroll(s32 difference);
