+++ common/source/tobkit/normalizebox.h
@@ -46,9 +46,9 @@ class NormalizeBox: public Widget {
 		void pleaseDraw(void);
 		
 		// Event calls
-		void penDown(u8 px, u8 py);
-		void penMove(u8 px, u8 py);
-		void penUp(u8 px, u8 py);
+		void penDown(u16 px, u16 py);
+		void penMove(u16 px, u16 py);
+		void penUp(u16 px, u16 py);
 		
 		s16 getValue(void);
