+++ common/source/tobkit/recordbox.h
@@ -51,8 +51,8 @@ class RecordBox: public Widget {
 		void pleaseDraw(void);
 		
 		// Event calls
-		void penDown(u8 px, u8 py);
-		void penUp(u8 px, u8 py);
+		void penDown(u16 px, u16 py);
+		void penUp(u16 px, u16 py);
 		void buttonPress(u16 button);
 		void buttonRelease(u16 button);
