+++ tobkit/include/tobkit/button.h
@@ -29,7 +29,7 @@ namespace tobkit {
 
 class Button: public Widget {
 	public:
-		Button(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen, bool _visible=true);
+		Button(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen, bool _visible=true);
 	
 		~Button();
 		
@@ -40,9 +40,9 @@ class Button: public Widget {
 		void pleaseDraw(void);
 		
 		// Event calls
-		void penDown(u8 x, u8 y);
-		void penUp(u8 x, u8 y);
-		void penMove(u8 x, u8 y);
+		void penDown(u16 x, u16 y);
+		void penUp(u16 x, u16 y);
+		void penMove(u16 x, u16 y);
 		void buttonPress(u16 button);
 		
 		inline bool isPenDown(void) const { return penIsDown; }
