+++ tobkit/include/tobkit/checkbox.h
@@ -24,7 +24,7 @@ namespace tobkit {
 class CheckBox: public Widget
 {
 	public:
-		CheckBox(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen,
+		CheckBox(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen,
 			 bool _visible=true, bool checked=false, bool albino=false);
 		~CheckBox();
 		
@@ -35,7 +35,7 @@ class CheckBox: public Widget
 		void pleaseDraw(void);
 		
 		// Event calls
-		void penDown(u8 px, u8 py);
+		void penDown(u16 px, u16 py);
 		
 		void registerToggleCallback(void (*onToggle_)(bool));
