+++ tobkit/include/tobkit/groupbox.h
@@ -24,7 +24,7 @@ namespace tobkit {
 class GroupBox: public Widget
 {
 	public:
-		GroupBox(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen, bool _visible=true);
+		GroupBox(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen, bool _visible=true);
 		~GroupBox();
 		
 		void pleaseDraw(void);
