+++ tobkit/include/tobkit/label.h
@@ -23,7 +23,7 @@ namespace tobkit {
 
 class Label: public Widget {
 	public:
-		Label(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen, bool _has_border=true,
+		Label(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen, bool _has_border=true,
 		      bool _albino=false, bool _no_bg=false, bool _right_aligned=false);
 		~Label(void);
 		
@@ -34,7 +34,7 @@ class Label: public Widget {
 		void pleaseDraw(void);
 		
 		// Event calls
-		void penDown(u8 x, u8 y);
+		void penDown(u16 x, u16 y);
 		
 		void setCaption(const char *caption);
 		char *getCaption(void);
