+++ tobkit/include/tobkit/listbox.h
@@ -36,16 +36,16 @@ namespace tobkit {
 
 class ListBox: public Widget {
 	public:
-		ListBox(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen, u16 n_items,
+		ListBox(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen, u16 n_items,
 			bool _show_numbers, bool _visible=true, bool _zero_offset=true);
 
 		// Drawing request
 		void pleaseDraw(void);
 
 		// Event calls
-		void penDown(u8 px, u8 py);
-		void penUp(u8 px, u8 py);
-		void penMove(u8 px, u8 py);
+		void penDown(u16 px, u16 py);
+		void penUp(u16 px, u16 py);
+		void penMove(u16 px, u16 py);
 
 		// Callback registration
 		void registerChangeCallback(void (*onChange_)(u16));
