+++ tobkit/include/tobkit/numberbox.h
@@ -23,14 +23,14 @@ namespace tobkit {
 
 class NumberBox: public Widget {
 	public:
-		NumberBox(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen, u8 _value=0, u8 _min=0, u8 _max=255, u8 _digits=2);
+		NumberBox(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen, u8 _value=0, u8 _min=0, u8 _max=255, u8 _digits=2);
 	
 		// Drawing request
 		void pleaseDraw(void);
 		
 		// Event calls
-		void penDown(u8 px, u8 py);
-		void penUp(u8 px, u8 py);
+		void penDown(u16 px, u16 py);
+		void penUp(u16 px, u16 py);
 
 		void setValue(u8 val);
