+++ tobkit/include/tobkit/pixmap.h
@@ -23,7 +23,7 @@ namespace tobkit {
 
 class Pixmap: public Widget {
 	public:
-		Pixmap(u8 _x, u8 _y, u8 _width, u8 _height, const u16* _image, Screen *_screen, bool _visible=true);
+		Pixmap(u16 _x, u16 _y, u16 _width, u16 _height, const u16* _image, Screen *_screen, bool _visible=true);
 	
 		~Pixmap();
 		
@@ -31,7 +31,7 @@ class Pixmap: public Widget {
 		void registerPushCallback(void (*onPush_)(void));	
 		
 		// Event calls
-		void penDown(u8 x, u8 y);
+		void penDown(u16 x, u16 y);
 		
 		// Drawing request
 		void pleaseDraw(void);
