+++ tobkit/include/tobkit/radiobutton.h
@@ -39,14 +39,14 @@ class RadioButton: public Widget {
 				void (*onChange)(u8);
 		};
 		
-		RadioButton(u8 _x, u8 _y, u8 _width, u8 _height, Screen *_screen,
+		RadioButton(u16 _x, u16 _y, u16 _width, u16 _height, Screen *_screen,
 				RadioButtonGroup *_rbg, bool _visible=true);
 		
 		// Drawing request
 		void pleaseDraw(void);
 		
 		// Event calls
-		void penDown(u8 px, u8 py);
+		void penDown(u16 px, u16 py);
 		
 		void setCaption(const char *caption);
 		void setActive(bool _active);
