+++ tobkit/source/gradienticon.cpp
@@ -22,7 +22,7 @@ using namespace tobkit;
 
 /* ===================== PUBLIC ===================== */
 
-GradientIcon::GradientIcon(u8 _x, u8 _y, u8 _width, u8 _height, const u32* _image, Screen *_screen, bool _visible)
+GradientIcon::GradientIcon(u16 _x, u16 _y, u16 _width, u16 _height, const u32* _image, Screen *_screen, bool _visible)
 	:Widget(_x, _y, _width, _height, _screen, _visible),
 	onPush(0), image(_image)
 {
@@ -40,7 +40,7 @@ void GradientIcon::registerPushCallback(void (*onPush_)(void)) {
 }
 
 // Event calls
-void GradientIcon::penDown(u8 x, u8 y) {
+void GradientIcon::penDown(u16 x, u16 y) {
 	if(onPush) onPush();
 }
